+++ CHANGES.rst
@@ -216,6 +216,9 @@ Miscellaneous
 
   - The ERFA wrappers are now written directly in the Python/C API
     rather than using Cython, for greater performance. [#3521]
+- Miscellaneous
+
+  - Improve import time of astropy [#3488].
 
 1.0.1 (unreleased)
 ------------------